Use hostname only to search for issue.
For incidents triggered by Nagios, only the hostname used to search
for the last updated issue. It would cause to open less duplicated
issues on our usecase. Configuration parameters can be added later
to make it more generalized.

Remove the optimization to avoid searching when issue key exists
on the subject field of the summary as it can only be used on trigger
action.