Do not filter incidents by status.
Log entries were filtered by the status field on the API configuration
by comparing with their incident status. The main purpose for this was
not to create issues for incidents already resolved. Move this logic
to the checker class and remove from the configuration.

This hard-to-understang configuration parameter used for this purpose
before to make the code completely unaware from the incident statuses
on the PagerDuty but the resolved status had to be checked before
trying to update incidents. So there are not any reason now to keep
that parameter on the configuration.